Restructure project to use Leiningen.
This involved adding a project.clj file and moving the textjure.clj
file into an appropriate directory structure and wrapping the naked
body of texture in a -main function. The switch over to -main required
that I, in turn, switch *command-line-args* over to use the arguments
passed in to -main.

Also, change deprecated reader macro "^" to be (meta ...).

Finally, I added a .gitignore to ignore the Leiningen lib/ directory.